FInally bringing all the shader compilation related fuinctions in the Backend in the same cpp, making specialized version based on the Backend
