+++ libraries/gl/src/gl/GLShaders.cpp
@@ -6,9 +6,9 @@ namespace gl {
 
 
 #ifdef SEPARATE_PROGRAM
-    bool compileShader(GLenum shaderDomain, const std::string& shaderSource, const std::string& defines, GLuint &shaderObject, GLuint &programObject) {
+    bool compileShader(GLenum shaderDomain, const std::string& shaderSource, const std::string& defines, GLuint &shaderObject, GLuint &programObject, std::string& error) {
 #else
-    bool compileShader(GLenum shaderDomain, const std::string& shaderSource, const std::string& defines, GLuint &shaderObject) {
+    bool compileShader(GLenum shaderDomain, const std::string& shaderSource, const std::string& defines, GLuint &shaderObject, std::string& error) {
 #endif
     if (shaderSource.empty()) {
         qCDebug(glLogging) << "GLShader::compileShader - no GLSL shader source code ? so failed to create";
@@ -69,6 +69,8 @@ namespace gl {
         }
         qCWarning(glLogging) << "GLShader::compileShader - errors:";
         qCWarning(glLogging) << temp;
+
+        error = std::string(temp);
         delete[] temp;
 
         glDeleteShader(glshader);
@@ -130,7 +132,7 @@ namespace gl {
     return true;
 }
 
-GLuint compileProgram(const std::vector<GLuint>& glshaders) {
+GLuint compileProgram(const std::vector<GLuint>& glshaders, std::string& error) {
     // A brand new program:
     GLuint glprogram = glCreateProgram();
     if (!glprogram) {
@@ -169,6 +171,9 @@ GLuint compileProgram(const std::vector<GLuint>& glshaders) {
 
         qCDebug(glLogging) << "GLShader::compileProgram -  failed to LINK the gl program object :";
         qCDebug(glLogging) << temp;
+
+        error = std::string(temp);
+        delete[] temp;
 
         /*
         filestream.open("debugshader.glsl.info.txt");
