+++ libraries/gpu-gl/src/gpu/gl41/GL41BackendBuffer.cpp
@@ -18,11 +18,17 @@ namespace gpu {
                 return result;
             }
 
+            ~GL41Buffer() { 
+                if (_texBuffer) {
+                    glDeleteTextures(1, &_texBuffer);
+                }
+            }
+
         public:
             GL41Buffer(const std::weak_ptr<gl::GLBackend>& backend, const Buffer& buffer, GL41Buffer* original) : Parent(backend, buffer, allocate()) {
-                glBindBuffer(GL_ARRAY_BUFFER, _buffer);
-                glBufferData(GL_ARRAY_BUFFER, _size, nullptr, GL_DYNAMIC_DRAW);
-                glBindBuffer(GL_ARRAY_BUFFER, 0);
+                glBindBuffer(GL_COPY_WRITE_BUFFER, _buffer);
+                glBufferData(GL_COPY_WRITE_BUFFER, _size, nullptr, GL_DYNAMIC_DRAW);
+                glBindBuffer(GL_COPY_WRITE_BUFFER, 0);
 
                 if (original && original->_size) {
                     glBindBuffer(GL_COPY_WRITE_BUFFER, _buffer);
@@ -36,19 +42,33 @@ namespace gpu {
             }
 
             void transfer() override {
-                glBindBuffer(GL_ARRAY_BUFFER, _buffer);
+                glBindBuffer(GL_COPY_WRITE_BUFFER, _buffer);
                 (void)CHECK_GL_ERROR();
                 Size offset;
                 Size size;
                 Size currentPage { 0 };
                 auto data = _gpuObject._renderSysmem.readData();
                 while (_gpuObject._renderPages.getNextTransferBlock(offset, size, currentPage)) {
-                    glBufferSubData(GL_ARRAY_BUFFER, offset, size, data + offset);
+                    glBufferSubData(GL_COPY_WRITE_BUFFER, offset, size, data + offset);
                     (void)CHECK_GL_ERROR();
                 }
-                glBindBuffer(GL_ARRAY_BUFFER, 0);
+                glBindBuffer(GL_COPY_WRITE_BUFFER, 0);
                 (void)CHECK_GL_ERROR();
                 _gpuObject._renderPages._flags &= ~PageManager::DIRTY;
+            }
+
+            // REsource BUffer are implemented with TextureBuffer 
+            GLuint _texBuffer { 0 };
+            GLuint getTexBufferId() {
+                if (!_texBuffer) {
+                    glGenTextures(1, &_texBuffer);
+                    glActiveTexture(GL_TEXTURE0 + GL41Backend::RESOURCE_BUFFER_TEXBUF_TEX_UNIT);
+                    glBindTexture(GL_TEXTURE_BUFFER, _texBuffer);
+                    glTexBuffer(GL_TEXTURE_BUFFER, GL_RGBA32F, _texBuffer);
+                    glBindTexture(GL_TEXTURE_BUFFER, 0);
+                    (void)CHECK_GL_ERROR();
+                }
+                return _texBuffer;
             }
         };
     }
@@ -63,14 +83,24 @@ GLuint GL41Backend::getBufferID(const Buffer& buffer) {
     return GL41Buffer::getId<GL41Buffer>(*this, buffer);
 }
 
+GLuint GL41Backend::getResourceBufferID(const Buffer& buffer) {
+    auto* object = GL41Buffer::sync<GL41Buffer>(*this, buffer);
+    if (object) {
+        return object->getTexBufferId();
+    } else {
+        return 0;
+    }
+}
+
 GLBuffer* GL41Backend::syncGPUObject(const Buffer& buffer) {
     return GL41Buffer::sync<GL41Buffer>(*this, buffer);
 }
 
 bool GL41Backend::bindResourceBuffer(uint32_t slot, BufferPointer& buffer) {
-    GLBuffer* object = syncGPUObject((*buffer));
-    if (object) {
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, slot, object->_id);
+    GLuint texBuffer = GL41Backend::getResourceBufferID((*buffer));
+    if (texBuffer) {
+        glActiveTexture(GL_TEXTURE0 + GL41Backend::RESOURCE_BUFFER_SLOT0_TEX_UNIT + slot); 
+        glBindTexture(GL_TEXTURE_BUFFER, texBuffer);
 
         (void)CHECK_GL_ERROR();
 
@@ -85,7 +115,8 @@ bool GL41Backend::bindResourceBuffer(uint32_t slot, BufferPointer& buffer) {
 void GL41Backend::releaseResourceBuffer(uint32_t slot) {
     auto& buf = _resource._buffers[slot];
     if (buf) {
-        glBindBufferBase(GL_SHADER_STORAGE_BUFFER, slot, 0);
+        glActiveTexture(GL_TEXTURE0 + GL41Backend::RESOURCE_BUFFER_SLOT0_TEX_UNIT + slot); 
+        glBindTexture(GL_TEXTURE_BUFFER, 0);
         buf.reset();
     }
 }
