+++ libraries/gpu-gl/src/gpu/gl45/GL45Backend.h
@@ -284,6 +284,11 @@ protected:
     // Output stage
     void do_blit(const Batch& batch, size_t paramOffset) override;
 
+    // Shader Stage
+    std::string getBackendShaderHeader() const override;
+    void makeProgramBindings(ShaderObject& shaderObject) override;
+    int makeResourceBufferSlots(GLuint glprogram, const Shader::BindingSet& slotBindings,Shader::SlotSet& resourceBuffers) override;
+
     // Texture Management Stage
     void initTextureManagementStage() override;
 };
