+++ libraries/gpu/src/gpu/Format.h
@@ -178,6 +178,7 @@ enum Semantic {
 
     UNIFORM,
     UNIFORM_BUFFER,
+    RESOURCE_BUFFER,
     SAMPLER,
     SAMPLER_MULTISAMPLE,
     SAMPLER_SHADOW,
