+++ libraries/procedural/src/procedural/ProceduralSkybox.cpp
@@ -59,7 +59,7 @@ void ProceduralSkybox::render(gpu::Batch& batch, const ViewFrustum& viewFrustum,
     auto& procedural = skybox._procedural;
     procedural.prepare(batch, glm::vec3(0), glm::vec3(1), glm::quat());
     auto textureSlot = procedural.getShader()->getTextures().findLocation("cubeMap");
-    auto bufferSlot = procedural.getShader()->getBuffers().findLocation("skyboxBuffer");
+    auto bufferSlot = procedural.getShader()->getUniformBuffers().findLocation("skyboxBuffer");
     skybox.prepare(batch, textureSlot, bufferSlot);
     batch.draw(gpu::TRIANGLE_STRIP, 4);
 }
