+++ libraries/render/src/render/drawItemBounds.slv
@@ -28,9 +28,24 @@ struct ItemBound {
     vec4 boundDim_s;
 };
 
+#if __VERSION__ == 450
 layout(std140) buffer ssbo0Buffer {
     ItemBound bounds[];
 };
+ItemBound getItemBound(int i) {
+    return bounds[i];
+}
+#else
+uniform samplerBuffer ssbo0Buffer;
+ItemBound getItemBound(int i) {
+    int offset = 2 * i;
+    ItemBound bound;
+    bound.id_boundPos = texelFetch(ssbo0Buffer, offset);
+    bound.boundDim_s = texelFetch(ssbo0Buffer, offset + 1);
+    return bound;
+}
+#endif
+
 
 
 out vec4 varColor;
@@ -67,7 +82,7 @@ void main(void) {
 
     vec4 cubeVec = UNIT_BOX[UNIT_BOX_LINE_INDICES[vertexID]];
 
-    ItemBound bound = bounds[boundID];
+    ItemBound bound = getItemBound(boundID);
     vec3 boundPos = bound.id_boundPos.yzw;
     vec3 boundDim = bound.boundDim_s.xyz;
